cache channels before updating server data, fixes #374
This fixes a bug where channels we don't have access to (because we can't Read Messages or Connect) are attempted to be resolved by `GET /channels/{cid}`, when in fact, we already have the data for these channels in GUILD_CREATE and can properly cache them from there.
For example, this would cause exceptions to be raised every time a user joins or leaves (voice state update) a voice channel that denies Connect to our client.

For more information on this behavior, see https://github.com/meew0/discordrb/issues/374#issuecomment-313839853